# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LunarBase API",
    "description": "A powerful backend-as-a-service API built with Rust, Axum, and Diesel ORM",
    "contact": {
      "name": "LunarBase Team",
      "email": "contact@lunarbase.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    },
    {
      "url": "https://api.lunarbase.dev",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/collections/{collection_name}/ownership/my-records": {
      "get": {
        "tags": [
          "Ownership"
        ],
        "summary": "Get records owned by the current user",
        "operationId": "get_my_owned_records",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to return",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of records to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owned records retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/collections/{collection_name}/ownership/stats": {
      "get": {
        "tags": [
          "Ownership"
        ],
        "summary": "Get ownership statistics for a collection",
        "operationId": "get_ownership_stats",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ownership statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions - Admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/collections/{collection_name}/ownership/users/{user_id}/records": {
      "get": {
        "tags": [
          "Ownership"
        ],
        "summary": "Get records owned by a specific user (admin only)",
        "operationId": "get_user_owned_records",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to return",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of records to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User owned records retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions - Admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/collections/{collection_name}/records/{record_id}": {
      "get": {
        "tags": [
          "Records"
        ],
        "operationId": "get_record",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "record_id",
            "in": "path",
            "description": "Record ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Record retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RecordResponse"
                }
              }
            }
          },
          "404": {
            "description": "Record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Records"
        ],
        "operationId": "update_record",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "record_id",
            "in": "path",
            "description": "Record ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Record updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Records"
        ],
        "operationId": "delete_record",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "record_id",
            "in": "path",
            "description": "Record ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Record deleted successfully"
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/collections/{collection_name}/records/{record_id}/ownership": {
      "get": {
        "tags": [
          "Ownership"
        ],
        "summary": "Check if current user owns a specific record",
        "operationId": "check_record_ownership",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "record_id",
            "in": "path",
            "description": "Record ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ownership status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/collections/{collection_name}/records/{record_id}/ownership/transfer": {
      "post": {
        "tags": [
          "Ownership"
        ],
        "summary": "Transfer ownership of a record",
        "operationId": "transfer_record_ownership",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "record_id",
            "in": "path",
            "description": "Record ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwnershipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ownership transferred successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/collections/{name}": {
      "get": {
        "tags": [
          "Collections"
        ],
        "summary": "Get collection by name",
        "operationId": "get_collection",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Collection name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CollectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_CollectionResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "name",
              "schema",
              "is_system",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "created_at": {
                "type": "string",
                "example": "2024-01-01 12:00:00"
              },
              "description": {
                "type": [
                  "string",
                  "null"
                ],
                "example": "Collection for storing product data"
              },
              "display_name": {
                "type": [
                  "string",
                  "null"
                ],
                "example": "Product Collection"
              },
              "id": {
                "type": "integer",
                "format": "int32",
                "example": 1
              },
              "is_system": {
                "type": "boolean",
                "example": false
              },
              "name": {
                "type": "string",
                "example": "products"
              },
              "schema": {
                "$ref": "#/components/schemas/CollectionSchema"
              },
              "updated_at": {
                "type": "string",
                "example": "2024-01-01 12:00:00"
              }
            }
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponse_RecordResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "collection_id",
              "data",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "collection_id": {
                "type": "string",
                "example": "products"
              },
              "created_at": {
                "type": "string",
                "example": "2024-01-01 12:00:00"
              },
              "data": {

              },
              "id": {
                "type": "string",
                "example": "1"
              },
              "updated_at": {
                "type": "string",
                "example": "2024-01-01 12:00:00"
              }
            }
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponse_Value": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {

          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "CollectionSchema": {
        "type": "object",
        "required": [
          "fields"
        ],
        "properties": {
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldDefinition"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "FieldDefinition": {
        "type": "object",
        "required": [
          "name",
          "field_type",
          "required"
        ],
        "properties": {
          "default_value": {

          },
          "field_type": {
            "$ref": "#/components/schemas/FieldType"
          },
          "name": {
            "type": "string",
            "example": "name"
          },
          "required": {
            "type": "boolean",
            "example": true
          },
          "validation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationRules"
              }
            ]
          }
        }
      },
      "FieldType": {
        "type": "string",
        "enum": [
          "text",
          "number",
          "boolean",
          "date",
          "email",
          "url",
          "json",
          "file",
          "relation"
        ]
      },
...
      "TransferOwnershipRequest": {
        "type": "object",
        "required": [
          "new_owner_id"
        ],
        "properties": {
          "new_owner_id": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UpdateRecordRequest": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {

          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}